done 3/4 ktra 2

+++ src/main/java/com/example/bai2/repositories/CategoryRepository.java
@@ -1,9 +1,7 @@
 package com.example.bai2.repositories;
 
+import com.example.bai2.models.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.bai2.models.Category;
-
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-    
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 }
